hmland: handle commands spanning packet-boundaries
Previously each command which crossed packet boundaries (because
there were many commands in a packet) was wrongly executed as
2 seperate commands. This is now fixed by reading from the
network until a \r or \n is seen and only then executing the
command.
Fixes crashes of the hmcfgusb when fhem reconnects and sends a
huge list of known devices ('+'-packets).